Moved robots.py into domain folder, removed subfolder
