+++ app/domain/references/routes.py
@@ -28,7 +28,7 @@ from app.domain.references.models.models import (
     Reference,
 )
 from app.domain.references.reference_service import ReferenceService
-from app.domain.robots.robots import Robots
+from app.domain.robots import Robots
 from app.persistence.sql.session import get_session
 from app.persistence.sql.uow import AsyncSqlUnitOfWork
