--- app/domain/robots/__init__.py
@@ -1,1 +1,0 @@
-"""Robots."""
